Implement network requests

+++ app/src/main/java/com/maxela/android/model/Main.kt
@@ -6,7 +6,7 @@ data class Main (
 
 	@SerializedName("temp") val temp : Double,
 	@SerializedName("feels_like") val feels_like : Double,
-	@SerializedName("temp_min") val temp_min : Int,
+	@SerializedName("temp_min") val temp_min : Double,
 	@SerializedName("temp_max") val temp_max : Double,
 	@SerializedName("pressure") val pressure : Int,
 	@SerializedName("humidity") val humidity : Int
